+++ R/pat_aggregate.R
@@ -5,7 +5,7 @@
 
 #' @title Aggregation statistics for PrupleAir time series
 #' 
-#' @param pat PurpleAir Timeseries "pat" object from \code{pat_loadLatest()}
+#' @param pat PurpleAir Timeseries "pat" object
 #' @param period The time period to average to. Can be "sec", "min", "hour", 
 #' "day", "DSTday", "week", "month", "quarter" or "year". A number can also
 #'  precede these options followed by a space (i.e. "2 day" or "37 min").